Move peers unit test helpers to separate files
