+++ test/common/helpers.js
@@ -1,0 +1,10 @@
+'use strict';
+
+module.exports = {
+
+	randomInt: function (min, max) {
+		min = Math.ceil(min);
+		max = Math.floor(max);
+		return Math.floor(Math.random() * (max - min + 1)) + min;
+	}
+};
